Added support for sorted sets:
Enable by adding a sorted_by key with two fields: type and field.

Type specifies the type of the data, and that influences how it ends
up ordered. The supported types are: Integer, String, Date and DateTime.

The field element specifies the field we use to order the set. You can use
two values, a string or a CodeRef. The string will be used as the key in
the data hash. The CodeRef should return the value to use.

Tweaks:

 * Schema type_spec upgrades the field setting to a CodeRef if a string
   was used;
 * fetch_set() returns the elements of the sorted set in order;
 * add_to_set() updates the field we need to sort the set.

Signed-off-by: Pedro Melo <melo@simplicidade.org>