Stop branding slaves with master set in create_into_set():
When using create_into_set, the slave object would receive an extra key,
the set name, with a hash with the master pk and type.

There is no attempt to keep this information up-to-date when we remove the
slave from the set, or other operations that might require such update,
so we just stop doing this branding.

Signed-off-by: Pedro Melo <melo@simplicidade.org>